refactoring for contact views
