+++ ContactBook/app/src/main/java/com/example/contactbook/ui/views/mainscreens/contacts/AddContactFragment.kt
@@ -1,4 +1,4 @@
-package com.example.contactbook.ui.views.Main.contacts
+package com.example.contactbook.ui.views.mainscreens.contacts
 
 import android.content.Context
 import android.os.Bundle
@@ -7,16 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import com.example.contactbook.R
-import com.example.contactbook.data.entities.Contact
-import com.example.contactbook.data.services.InputValidationService
-import com.example.contactbook.ui.viewModels.ContactViewModel
-import com.example.contactbook.databinding.FragmentAddContactBinding
 import com.example.contactbook.data.services.AuthorisedSharedPreferencesService
 import com.example.contactbook.data.services.abstractions.IAuthorisedSharedPreferencesService
-import com.example.contactbook.data.services.abstractions.IInputValidationService
+import com.example.contactbook.databinding.FragmentAddContactBinding
+import com.example.contactbook.ui.viewModels.ContactViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import java.util.*
