+++ ContactBook/app/src/main/java/com/example/contactbook/ui/views/mainscreens/contacts/ContactsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.contactbook.ui.views.Main.contacts
+package com.example.contactbook.ui.views.mainscreens.contacts
 
 import android.content.Context
 import android.os.Bundle
@@ -7,8 +7,6 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
@@ -17,6 +15,7 @@ import com.example.contactbook.ui.viewModels.ContactViewModel
 import com.example.contactbook.databinding.FragmentContactsBinding
 import com.example.contactbook.data.services.AuthorisedSharedPreferencesService
 import com.example.contactbook.data.services.abstractions.IAuthorisedSharedPreferencesService
+import com.example.contactbook.ui.views.mainscreens.contacts.ContactsAdapter
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -26,7 +25,6 @@ class ContactsFragment : Fragment() {
 
     private val mContactViewModel : ContactViewModel  by viewModels()
 
-    private lateinit var adapter: ContactsAdapter
     private lateinit var contactsRecyclerView : RecyclerView
 
     override fun onCreateView(
@@ -36,7 +34,8 @@ class ContactsFragment : Fragment() {
         binding = FragmentContactsBinding.inflate(inflater, container, false)
 
         contactsRecyclerView = binding.contactListRecyclerView
-        adapter = ContactsAdapter()
+
+        val adapter = ContactsAdapter()
 
         contactsRecyclerView.adapter = adapter
         contactsRecyclerView.layoutManager = LinearLayoutManager(requireContext())
