+++ ContactBook/app/src/main/java/com/example/contactbook/ui/views/mainscreens/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.example.contactbook.ui.views.Main.settings
+package com.example.contactbook.ui.views.mainscreens.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.contactbook.databinding.FragmentSettingsBinding
+
 
 class SettingsFragment : Fragment() {
     private var _binding: FragmentSettingsBinding? = null
