+++ ContactBook/app/src/main/res/drawable/bg_rounded_tv.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android">
+    <corners
+        android:bottomRightRadius="45dp"
+        android:topRightRadius="45dp"
+        />
+    <stroke
+        android:width="1dp"
+        android:color="@color/additional"
+        />
+    <solid
+        android:color="@color/white"
+        />
+</shape>
