+++ ContactBook/app/src/main/res/values/themes.xml
@@ -2,8 +2,8 @@
     <!-- Base application theme. -->
     <style name="Theme.ContactBook" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/additional</item>
-        <item name="colorPrimaryVariant">@color/main</item>
+        <item name="colorPrimary">@color/main</item>
+        <item name="colorPrimaryVariant">@color/additional</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">@color/teal_200</item>
